backend/requirements/base.txt
# Core Django and DRF
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
dj-database-url==2.1.0
psycopg2-binary==2.9.9

# Configuration
python-decouple==3.8

# API Documentation
drf-yasg==1.21.7

# CORS handling
django-cors-headers==4.3.1

# Filtering
django-filter==23.5

# Password hashing
bcrypt==4.1.2

# Timezone support
pytz==2024.1

# Redis (for caching and rate limiting)
django-redis==5.4.0
redis==5.0.1

# Static files
whitenoise==6.6.0

# Image processing (for profile pictures)
Pillow==10.2.0

# Utilities
python-dateutil==2.8.2

# Error tracking and monitoring
sentry-sdk==1.40.0
